Update to API 6.1

+++ include/tgbot/types/Animation.h
@@ -60,9 +60,12 @@ public:
     std::string mimeType;
 
     /**
-     * @brief Optional. File size
+     * @brief Optional. File size in bytes.
+     * 
+     * It can be bigger than 2^31 and some programming languages may have difficulty/silent defects in interpreting it.
+     * But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this value.
      */
-    std::int32_t fileSize;
+    std::int64_t fileSize;
 };
 }
